hey, the setup generator actually works! still need to write tests
