+++ lib/railsmdb/cli.rb
@@ -9,4 +9,8 @@ require 'railsmdb/ext/rails/generators/rails/app/app_generator'
 Rails::AppLoader.send :remove_const, :EXECUTABLES
 Rails::AppLoader::EXECUTABLES = %w[ bin/railsmdb ].freeze
 
-require 'rails/cli'
+if ARGV.first == 'setup'
+  Rails::Command.invoke :setup, ARGV
+else
+  require 'rails/cli'
+end
