+++ lib/railsmdb/generators/setup/templates/README.md.tt
