+++ lib/railsmdb/generators/setup/templates/_bin/railsmdb.tt
