+++ lib/railsmdb/generators/setup/templates/_config/initializers/mongoid.rb.tt
